Cleaning data set
I dropped rows for which all values for a particular question were null. After this, I changed all null values to '0'.

In particular, for question set S15r1-11, leaving all entries blank resulted in that row being dropped. If a respondent answered at least one of S15r1-11 with a value greater than or equal to 1, then other null values for that row were changed to 0.

I also added a column labeled as 'Fan_mangitude' which took the magnitude of the vector of columns S15r1-11.